backend/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
pymongo==4.7.3
openai==1.30.1
anthropic==0.27.0
google-generativeai==0.5.4
requests==2.32.3
# Add other dependencies as needed, e.g., for Deepseek if a specific SDK is used
httpx==0.26.0
pydantic==2.7.4 